requirements.txt
Flask>=2.0
requests>=2.25
python-whois>=0.7
reportlab>=3.6
stem>=1.8
python-dotenv>=0.21
phonenumbers>=8.12
werkzeug>=2.0
